refactor(src): rename util to lib
- rename `util` to `lib`
- move Base Accessory class to lib
