+++ src/lib/dingzTypes.ts
