+++ src/lib/libs.d.ts
