+++ src/lib/myStromTypes.ts
